+++ Source/labyrinths/debugdungeon.js
@@ -167,9 +167,9 @@ module.exports = new Labyrinth("Debug Dungeon", "Untyped", 10, [10])
 		})
 	.setRooms(
 		{
-			"Event": ["Twin Pedestals", "Elemental Research", "Free Gold?", "Health Redistribution", "The Score Beggar", "Repair Kit, just hanging out", "Abandoned Forge", "Equipment Merchant", "Consumable Merchant", "Rest Site", "Treasure!"],
+			"Event": ["Twin Pedestals", "Elemental Research", "Free Gold?", "Health Redistribution", "The Score Beggar", "Repair Kit, just hanging out", "Abandoned Forge", "Equipment Merchant", "Consumable Merchant", "Overpriced Merchant", "Rest Site", "Treasure!"],
 			"Battle": ["Hawk Fight", "Frog Fight", "Mechabee Fight", "Slime Fight", "Tortoise Fight"],
-			"Merchant": ["Equipment Merchant", "Consumable Merchant"],
+			"Merchant": ["Equipment Merchant", "Consumable Merchant", "Overpriced Merchant"],
 			"Rest Site": ["Rest Site"],
 			"Final Battle": ["A Northern Laboratory", "Hall of Mirrors", "The Hexagon"],
 			"Forge": ["Abandoned Forge"],